flui-types: checked RGBA8 length in Image::from_rgba8; add try_from_rgba8 (#1288)
* flui-types: checked RGBA8 length in Image::from_rgba8; add try_from_rgba8

The expected length width * height * 4 was computed in u32: it panicked
on overflow in debug and wrapped in release, so a 65536x65536 image
matched an empty buffer. Compute it in usize with checked arithmetic.

Add a fallible Image::try_from_rgba8 returning ImageDataError, and use
it in the decoder paths (flui-assets ImageAsset, flui-widgets
decode_bytes). Remove from_rgba8_unchecked: no callers, and its Safety
section documented a safe fn.


* flui-engine: checked offsets in nine-slice sub-image extraction

The nine-slice region copy computed row offsets as u32 (`row * stride +
sx * 4`, `sy + sh`, `sw * sh * 4`): a large region wrapped in release
into an in-bounds read of the wrong pixels, or panicked in debug. Move
it to extract_rgba8_region, which does the arithmetic in checked usize,
reads rows through slice::get, and builds the result with
Image::try_from_rgba8.

The flui-widgets test helpers that sized RGBA8 buffers as
(width * height * 4) as usize now multiply in usize too.


---------

+++ crates/flui-widgets/src/image/decode_cache.rs
@@ -212,7 +212,11 @@ mod tests {
     }
 
     fn solid(width: u32, height: u32) -> PixelImage {
-        PixelImage::from_rgba8(width, height, vec![0u8; (width * height * 4) as usize])
+        PixelImage::from_rgba8(
+            width,
+            height,
+            vec![0u8; width as usize * height as usize * 4],
+        )
     }
 
     /// A cold key (never inserted by any test) is a guaranteed miss.
